The purpose of this data mining project was to perform classification using Support Vector Machine (SVM), Naïve Bayes, and K-Nearest Neighbors (KNN) algorithms
The purpose of this data mining project was to perform classification using Support Vector Machine (SVM), Naïve Bayes, and K-Nearest Neighbors (KNN) algorithms, replacing some values in the dataset with NaN and displaying the result of the replacement, Replace the NaN values you obtained in step 2 with mean of the corresponding column. Display the result of replacement, And Classification on the dataset you obtained in Step 3 with SVM, Naïve Bayes and KNN. Display confusion matrix. The project involved loading a dataset, preprocessing the data, splitting it into training and testing sets, applying feature scaling, training the classifiers, and making predictions on new data points.